% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\name{codon_checker}
\alias{codon_checker}
\title{Codon Usage Comparison Function}
\usage{
codon_checker(reference_codons, mutant_codons, species)
}
\arguments{
\item{reference_codons}{list of reference codons from sequence alignments}

\item{mutant_codons}{list of corresponding mutated codons from sequence alignments}

\item{species}{"staph" or "e.coli", specifies which codon usage table will be used}
}
\value{
summary table comparing usage frequencies of reference codons to corresponding mutated codons
}
\description{
This function assists in determining if synonymous mutations are likely silent
}
