% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{send_for_sequencing}
\alias{send_for_sequencing}
\title{Send for Sequencing Function}
\usage{
send_for_sequencing(method, length = 0, concentration)
}
\arguments{
\item{method}{"PCR" or "plasmid", refers to type of sample being prepared}

\item{length}{length of amplicon, required for PCR samples. Defaults to 0}

\item{concentration}{DNA concentration of samples, in ng/uL}
}
\value{
table with volumes of DNA, water, and primer required for sequencing
}
\description{
This function eliminates manual calculations related to sample preparation of plasmid or PCR product sequencing
}
