% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_plate.R
\name{tabulate_plate}
\alias{tabulate_plate}
\title{Tabulate Plate Function}
\usage{
tabulate_plate(file, identifiers, group_by, data)
}
\arguments{
\item{file}{path to file with microarray data and layout in plateR format}

\item{identifiers}{vector of column names that uniquely identify observations/treatment groups}

\item{group_by}{column containing variable to group data by e.g. replicate number}

\item{data}{vector of columns containing data output from the plate reader}
}
\value{
modified tabular data frame that groups data according to specifications
}
\description{
This function expands upon the functionality of plateR to easily tabulate 96-well format data
}
